Implemented season struct
Originally, Game struct was supposed to hold a reference to a &Team
struct. However, this was causing numerous problems with lifetimes when
trying to tie both the Game and Team to a single parent Season struct.
As such, I've eliminated all references and used clone() wherever
possible to enable quicker prototyping. If performance is slow, this
will need to be changed in the future.